pyproject.toml
[tool.poetry]
name = "miqro"
version = "1.2.0"
description = "MIQRO is an MQTT Micro-Service Library for Python"
authors = ["Daniel Fett <mail@danielfett.de>"]

[tool.poetry.dependencies]
python = "^3.7"
PyYAML = "^6.0"
paho-mqtt = "^1.6.1"
pydantic = "^1.10.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
